fix: no more static `World` (see long commit msg)
There was a state management bug with the `World` object where mutations
to the rules or template would carry over to other tests despite calling
the `reset()` function. In the process of debugging this I changed the
`World` class to be completely non-static and just use the `getWorld()`
singleton. This isn't required for the bug fix but I think it's more
hygienic code.