[DEV-3295] Various program test suite fixes (#296)
* fix: fix type error in "current time" step def

* refactor: null coalescence operator

* fix: fix broken user segment step step definitions

* fix: add missing program triggers to trigger step

Don't know why the diff for this is so funky, all I did was refactor the
regex a bit. Oh well.

* fix: no more static `World` (see long commit msg)

There was a state management bug with the `World` object where mutations
to the rules or template would carry over to other tests despite calling
the `reset()` function. In the process of debugging this I changed the
`World` class to be completely non-static and just use the `getWorld()`
singleton. This isn't required for the bug fix but I think it's more
hygienic code.

* chore: make defaultIntrospection private

* feat: add `dev` script command

* chore: organize imports

* feat: improve assertion failure error messages

* chore: release version 2.1.2-1

* chore: add CHANGELOG and README

* chore: publish version 2.2.0

* chore: publish program-test-suite version 2.2.1